+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Add `after_running_test_group` and `before_running_test_group` events
 - Add `Petitest::TestGroup#sub_test_group`
 - Improve error report format
 - Improve default `#assert` message